fix: popularity being integer instead of float

+++ src/models/torrent.rs
@@ -76,7 +76,7 @@ pub struct Torrent {
     /// Number of seeds connected to
     pub num_seeds: i64,
     /// Popularity of the torrent
-    pub popularity: i64,
+    pub popularity: f64,
     /// Torrent priority. Returns -1 if queuing is disabled or torrent is in seed mode
     pub priority: i64,
     /// True if torrent is from a private tracker (added in 5.0.0)
@@ -200,7 +200,7 @@ impl<'de> Visitor<'de> for TorrentMapVisitor {
             num_incomplete: i64,
             num_leechs: i64,
             num_seeds: i64,
-            popularity: i64,
+            popularity: f64,
             priority: i64,
             private: Option<bool>,
             progress: f32,
